Фикс редактирования истории

+++ app/src/main/java/pro/progr/todos/EditCardInHistoryViewModel.kt
@@ -86,7 +86,7 @@ class EditCardInHistoryViewModel private constructor(
             viewModelScope.launch(Dispatchers.Default) {
                 repository.updateCardInHistory(title = _cardContent.value.title,
                     description = _cardContent.value.text,
-                    noteId = cardId.toLong(),
+                    noteId = cardId,
                     epochDay = epochDay)
             }
